Final Comit
I added two more data types to make the comparison of ints more broad. This includes list which is already committed and  a HashSet. I also added two more test cases in the Program.cs file. There was also a LinkedList missing from the code so I added it back in.